phpStan cleanup & fix handleOptionsDefined().
There was a significant bug where handleOptionsDefined() did not
match the behavior described in its docblock, which was its
intended behavior. The actual behavior has been moved to a new
method called handleOptionsAllowed() and the handleOptionsDefined()
has been rewritten to match the behavior described in the
docblock.

This is a breaking change.

Deprecated methods have been removed.